Move surface_id to a common variable.
Debugging FreeRDP when connecting to the The Windows 10
RDP server shows Windows  defaults to egfx channel 0.

Thus, we should have parity with that.

However, I found changing that value difficult because it's repeated in
so many places. This modification both moves the value to 0 and makes it
easy to change or refactor should we chose to do so later.